Add disconnect functionality
