+++ pages/_app.tsx
@@ -1,5 +1,7 @@
 import type { AppProps } from "next/app";
 import { useEffect } from "react";
+import { Web3ReactProvider } from "@web3-react/core";
+import { Web3Provider } from "@ethersproject/providers";
 import { ThemeProvider, CssBaseline } from "@mui/material";
 import theme from "../theme";
 
@@ -12,11 +14,19 @@ function MyApp({ Component, pageProps }: AppProps) {
     }
   }, []);
 
+  const getLibrary = (provider: any): Web3Provider => {
+    const library = new Web3Provider(provider);
+    library.pollingInterval = 12000;
+    return library;
+  };
+
   return (
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      <Component {...pageProps} />
-    </ThemeProvider>
+    <Web3ReactProvider getLibrary={getLibrary}>
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        <Component {...pageProps} />
+      </ThemeProvider>
+    </Web3ReactProvider>
   );
 }
